payment-recovery-system-extraction/requirements.txt
# Invoice Extraction Service Dependencies

# FastAPI and Web Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# HTTP Client for Java API
requests==2.31.0
httpx==0.26.0

# PDF Processing
pdfplumber==0.10.3

# Image Processing
Pillow==10.2.0
pytesseract==0.3.10

# Excel Processing
pandas==2.2.0
openpyxl==3.1.2
xlrd==2.0.1

# Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Logging
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

